ocfemia-lizelle-2023-hw6/requirements.txt
matplotlib
pytest
pyyaml
tqdm
scikit-learn
datasets
einops